Added pineapple.easypatch
Accepts a raws file as input and, where applicable, adds PERMITTED_X tokens and related to entities. Simple but should be really very handy for porting the less complicated mods very quickly. And in most cases an end user could just give easypatch a path to some raws file or directory of them to add and everything will just work.